pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ANIME-RECOMMENDER"
version = "0.1"
authors = [
    {name = "Hemanth"}
]
description = "AI-powered anime recommendation system"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "langchain",
    "langchain-community",
    "langchain_groq",
    "chromadb",
    "streamlit",
    "pandas",
    "python-dotenv",
    "sentence-transformers",
    "langchain_huggingface",
]

[tool.setuptools]
packages = {find = {}}
